requirements.txt
# requirements.txt - will expand later if needed
pyyaml
playwright
openai
requests
transformers
torch